requirements.txt
tensorflow
opencv-python
numpy
keras
supabase
python-dotenv
scipy
customtkinter
requests
Pillow
